Fix the data-ingestion-artifacts references
I have seen in the readme for fresh start there is `upload-artifacts` function mentioned but I found it confusing to apply additional steps when we are already creating resources through iac terraform code. And also to provide consistency with data-transformation iac code I am suggesting the change. But not sure if seperating artifacts in different bucket has additional advantages.